% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurora.R
\name{get_macroeconomic_data}
\alias{get_macroeconomic_data}
\title{Get Macroeconomic Data}
\usage{
get_macroeconomic_data(start_date, end_date, latitude, longitude)
}
\arguments{
\item{start_date}{Character string. Start date.}

\item{end_date}{Character string. End date.}

\item{latitude}{Numeric. Latitude.}

\item{longitude}{Numeric. Longitude.}
}
\value{
A merged data frame containing all indicators.
}
\description{
Orchestrates the fetching and merging of weather, economic, and political/disaster data.
}
