% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurora.R
\name{gdelt_timeline_daily}
\alias{gdelt_timeline_daily}
\title{Get GDELT Timeline Data}
\usage{
gdelt_timeline_daily(query_text, out_col, start, end)
}
\arguments{
\item{query_text}{Character string. Query text for GDELT.}

\item{out_col}{Character string. Name of the output column for the score.}

\item{start}{Character string. Start date.}

\item{end}{Character string. End date.}
}
\value{
A data frame with Date and the requested score column.
}
\description{
Fetches sentiment tone timeline from GDELT API.
}
