% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurora.R
\name{gdelt_timeline_daily}
\alias{gdelt_timeline_daily}
\title{Get a daily GDELT TimelineTone score and align to a date spine}
\usage{
gdelt_timeline_daily(query_text, out_col, start, end)
}
\arguments{
\item{query_text}{Character. GDELT query string.}

\item{out_col}{Character. Name of the score column to create.}

\item{start}{Character. Start date in "YYYY-MM-DD".}

\item{end}{Character. End date in "YYYY-MM-DD".}
}
\value{
A data.frame with columns: date and the score column.
}
\description{
Get a daily GDELT TimelineTone score and align to a date spine
}
