% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurora.R
\name{plotter}
\alias{plotter}
\title{Plot revenue vs selected variables with scaling}
\usage{
plotter(
  df,
  cols,
  date_col = "Date",
  revenue_col = "Revenue",
  ncol = NULL,
  save_path = NULL,
  width = 14,
  height = 9,
  dpi = 300,
  scale_method = c("zscore", "minmax"),
  drop_na = TRUE
)
}
\description{
Plot revenue vs selected variables with scaling
}
